# web/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Discover hidden world of palindromes",
        "title": "PalindromeX",
        "contact": {},
        "license": {
            "name": "BSD 2-Clause License",
            "url": "https://choosealicense.com/licenses/bsd-2-clause/"
        },
        "version": "1.0"
    },
    "host": "palindromex.ml",
    "basePath": "/v1/users/{userID}",
    "paths": {
        "/messages": {
            "get": {
                "security": [
                    {
                        "ApiToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves messages that belong to a specified user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Message"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiToken": []
                    }
                ],
                "consumes": [
                    "text/plain"
                ],
                "summary": "Updates existing message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "messageID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message (palindrome text)",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {}
                }
            },
            "post": {
                "security": [
                    {
                        "ApiToken": []
                    }
                ],
                "consumes": [
                    "text/plain"
                ],
                "summary": "Creates a new message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message (palindrome text)",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {}
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiToken": []
                    }
                ],
                "summary": "Deletes existing message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "messageID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {},
                    "404": {}
                }
            }
        },
        "/messages/{mesageID}": {
            "get": {
                "security": [
                    {
                        "ApiToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves one message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "messageID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "palindrome": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}